+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/react
 
+## 0.0.174
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/toast@0.0.25
+
 ## 0.0.173
 
 ### Patch Changes